rays: Extract OpenGL-specific code from painter.cpp to opengl/painter.cpp
Split painter implementation into GL-free (src/painter.cpp) and
GL-specific (src/opengl/painter.cpp) parts. PainterData inherits from
Painter::Data for polymorphic GL state management. Move draw helpers
to polygon.cpp as file-local draw_polygon functions.
